requirements-dev.txt
Pillow
nose
